Rename src/lib/simon to src/lib/game and strip simon prefix from file names #256
