+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "mnemecha",
-	"version": "0.111.0",
+	"version": "0.112.0",
 	"type": "module",
 	"private": true,
 	"scripts": {
