+++ src/lib/game/HardScene.svelte.spec.ts
@@ -1,12 +1,12 @@
 import type { ScoreData } from '@joshuafolkken/game-kit'
 import { describe, expect, it, vi } from 'vitest'
 import { render } from 'vitest-browser-svelte'
-import HardSimonScene from './HardSimonScene.svelte'
+import HardSimonScene from './HardScene.svelte'
 import type { HardSimonBoardData } from './types'
 
 vi.mock('@threlte/core', () => ({ T: {}, useTask: vi.fn() }))
 vi.mock('@threlte/extras', () => ({ Text: function Text() {} }))
-vi.mock('./HardSimonBoard.svelte', () => ({ default: function HardSimonBoard() {} }))
+vi.mock('./HardBoard.svelte', () => ({ default: function HardSimonBoard() {} }))
 vi.mock('@joshuafolkken/game-kit', () => ({ ScoreDisplay: function ScoreDisplay() {} }))
 vi.mock('./hard-board-config', () => ({
 	HARD_BOARD_X_LEFT: -1.95,
