+++ src/lib/game/board-center-label.ts
@@ -4,7 +4,7 @@ import {
 	MULTILINE_LINE_HEIGHT,
 	ROUND_DIGIT_FONT_SIZE,
 	SINGLE_LINE_HEIGHT,
-} from './simon-button-config'
+} from './button-config'
 
 interface BoardCenterLabelInput {
 	phase: string
