+++ src/lib/game/board-config.spec.ts
