+++ src/lib/game/board-config.ts
