+++ src/lib/game/board-input.ts
