+++ src/lib/game/button-config.ts
