+++ src/lib/game/credits.spec.ts
