+++ src/lib/game/credits.ts
