+++ src/lib/game/flash.ts
