+++ src/lib/game/hard-board-config.ts
