+++ src/lib/game/hard-board-input.ts
