+++ src/lib/game/score.svelte.ts
